+++ src/main/java/step2/LotteryMain.java
@@ -1,5 +1,8 @@
 package step2;
 
+import step2.view.InputView;
+import step2.view.ResultView;
+
 public class LotteryMain {
     public static void main(String[] args) {
         LotteryController lotteryController = new LotteryController(new InputView(), new ResultView());
