+++ src/main/java/step2/view/InputView.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.view;
 
 import step2.domain.LotteryNumber;
 import step2.domain.Money;
@@ -6,7 +6,7 @@ import step2.domain.Money;
 import java.util.Arrays;
 import java.util.Scanner;
 
-class InputView {
+public class InputView {
     public Money requestMoney() {
         System.out.println("구입금액을 입력해 주세요.");
         Scanner scanner = new Scanner(System.in);
