+++ src/test/java/step2/LotteryControllerTest.java
@@ -3,6 +3,8 @@ package step2;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import step2.domain.*;
+import step2.view.InputView;
+import step2.view.ResultView;
 
 import java.util.ArrayList;
 import java.util.List;
